refactor(gridSession): move grid session handling to express
